Makefile: add support for in.tftpd
The atftpd package is not available on fedora.  The default tftpd
package is in.tftpd.  I have been using this and it works fine, although
the logs may not be as nice.

Allow the make file to fall back to use in.tftpd if atftpd is not
available.  Report an error if nothing is available.