Make it possible to start an image in qemu as a user.
By switching to usermode networking in qemu, because of this we no
longer need to set up the tap interface, nor do we need to start tftpd
as root.

This change requires a fix to the netboot code in the bios, otherwise it
will time out fetching the second block of the firmware (the second 512
bytes).