TFTP: Support alternate TFTP Server Port
If TFTP_SERVER_PORT is set in the environment, TFTP server will be
started on that port; if not, it is started on UDP/69 (well known
TFTP service port) as before.

If TFTP_SERVER_PORT is set in environment during build of litex
BIOS, then TFTP client will try to reach TFTP server on that port
first before falling back to UDP/69.

If TFTP_SERVER_PORT >= 1024 then TFTP server is started as user
without using sudo (otherwise sudo is used as before).

Tested with atftpd; in.tftpd equivalent added but not tested.