firmware: stub-out framebuffer/pattern if no main ram
On systems without external SDRAM (eg, Numato Elbert v2), there
is no MAIN_RAM_BASE, and we cannot hold a frame buffer as we
have only small amounts of in-FPGA SRAM.

This should be cleaned up to avoid even compiling the framebuffer /
pattern when there is no main ram / MAIN_RAM_BASE, but for now we
have stubbed out everything referring to main ram or L2 cache (since
the L2 cache is over the external SDRAM).

In theory this should still work on systems with MAIN RAM, but currently
untested.