firmware/linker.ld: WIP: hand-calculated Elbert v2 values
UGLY!  Hand hacked linker.ld to allow linking on platform without
main_ram.

TODO: figure out way to generate this conditional on whether there
is main ram or not.

On platforms with main ram, we tell linker to assume that most things
will be in main_ram by the time we execute the code (and the boot
loader copies into the main ram).  On a platform without main ram,
we have to execute directly from the SPI flash (which is mapped
in a different location), and need to:
(a) ensure that the .text (code) and .rodata (const data) is in
the spi flash; and

(b) .data (updatable data) and .bss (stack) are in sram (static ram,
on FPGA cells)