Mimas v2: make gateware-flash include BIOS
On the Mimas v2 the Gateware and the BIOS need to be stored in the same
flash on the Mimas v2 board, updated via the same method (MimasV2Config.py).
As a result it is not very usfeul to *just* install the Gateware, with
no BIOS image, as the only way to get a BIOS image onto the board would
be to use "make image-flash" with no application, which would also load
the Gateware again.

Redirect the "make gateware-flash" on the Mimas v2 to load an image
that contains the Gateware + BIOS, but no application.  The result is:

    make gateware-flash - flashes Gateware+BIOS, make firmware-load works
    make image-flash    - flashes Gateware+BIOS+Firmware, use firmware-connect