qemu: Fix spiflash args as per deprecated args
The arg available to specify serial for -drive was to be used for
specifying a drive serial number.  However, we we using it to specify
the flash driver.  This was a hack.

Introduce a driver option to be able to specify this on litex_ssi.  This
will then be read during ssi bus / drive initialization to initialize
the requested guest hardware.

By default we fall back to m25p80.

Example usage:
  -drive if=mtd,format=qcow2,file=build/arty_net_or1k//qemu.qcow2
  -global litex_ssi.spiflash=mx25l1606e