qemu: Allow qemu build to get upstream qemu after changes
Currently the qemu build will checkout the requested qemu branch if we
are not on it.  However, if we are on the correct branch it will not do
anything.  This is usually fine, but if upstream has changes it means
those changes will be be able to get picked up.

Add a 'git reset' to allow picking up of the upstream changes.

Also, currently the QEMU_REMOTE_NAME is always the same, that means if
we override the remote we will not really be able to add the new remote.
To enable this create the QEMU_REMOTE_NAME based on the remote url.