litex: sim: Get sim working in litex-buildenv
Changes needed to get PLATFORM=sim to work again.

The platforms/sim was basically copied from litex.boards.platforms.sim
but then updated to properly handle sending the SimConfig details to the
toolchain needed during building and running of the verilog backend.

The changes in gateware/firmware are needed to allow the firmware to be
loaded lazily at build time.  This is because during build we use a
single SoC for firmware and gateware build.  Since the old code was
looking for the firmware during SoC construction it didnt work.