firmware: Detect command on \n consistent with telnet
Some uart/terminals are sending only '\n'.  I think it makes sense
that we detect commands when a '\n' comes in and ignore the '\r'
this is consistent with the telnet branch.

This fixes an issue in verilator where the serail2console module
sometimes doesnt send '\n' when we hit <enter>.