scripts: Set CONDA_PREFIX in enter-env.sh and download-env.sh
This is required for Renode package to run properly. This variable is
set automatically when activating an environment, but not if we work
with the base one, without activation.

Setting it in download-env will ensure the proper path when running it
from an activated environment. Please note this is not a default
usage scenario.