linux: unify linux source
This commit changes the linux source
for VexRiscv and mor1kx to litex-hub/linux.

Since this is the same repository as used
by linux-on-litex-vexriscv and all the LiteX
drivers are included, there is no more need for
additional BUILD_BUILDROOT option.

This also updates rootfs images for both architectures
and adds buildroot configs used to generate them.